release 0.1.202 source code

+++ services/gaussdb/v3/model/model_backups.go
@@ -33,7 +33,7 @@ type Backups struct {
 	Type *string `json:"type,omitempty"`
 
 	// 备份大小（单位：MB）。
-	Size *int64 `json:"size,omitempty"`
+	Size *float64 `json:"size,omitempty"`
 
 	Datastore *MysqlDatastoreInBackup `json:"datastore,omitempty"`
